fix(process): Passing all environment variables
The original code updated the envvars during execution. This was later
changed to only pass the debug flag and the tcp server flag, however
this breaks the node process as libuv requires `HOME` envvar to be
present (and who knows what else). For doing the right thing here and
fot future proofing this, we now copy all envvars and add the ones which
we explicitly control via application parameters.