feat(agent): Sanitize messages for chat restore
When restory a chat session, Cody expects a list of dictionaries which
contains two keys: "text" and "speaker". Previously we passed Message
objects which also implements these fields however it also contained
contextFiles. If we just pass these, the contextField field will eat up
tokens and trigger context limit violations. Also this field is
technically irrelevant for the conversation thread so we're ensuring
it's not sent while restoring a session.