fix(messaging): Ensure only one coroutine can read/write socket
Since this is asynchronous code, in some truly async application use it
happened that multiple coroutines tried to read/write the socket the
same time. This is of course not possible if the socket hasn't finished
sending data. For making this concurrency safe, we're using a lock to
ensure only a single coroutine can read/write at a given time, and block
the others until it's done.