Update to the latest Cody
Previously, Codypy was using the old standalone binary distribution of
Cody. This was problematic because this distribution is no longer
supported and it crashes when using the latest version of Sourcegraph.

This PR updates the code to use the latest supported version of the Cody
JSON-RPC protocol. Instead of the old standalone binary, Cody is now
only distributed as an `index.js` bundle that is published to npm.
This PR downloads the index.js file directly from npm (via tarball
download) and creates a script that runs `node package dist/index.js`.
This PR also updates the code to fix some Pydantic crashes that are
caused by assuming some fields are required when they are optional.